♻️ refactor(mobile/app): update import paths to use layers for better organization
